리셋버튼 스크립트
TileGenerator restart가 reset과 중복이여서 삭제했습니다
