+++ Assets/Scenes/MainScene.unity
@@ -2400,7 +2400,7 @@ GameObject:
   - component: {fileID: 919837663}
   - component: {fileID: 919837662}
   m_Layer: 5
-  m_Name: Button
+  m_Name: Reset
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
