+++ Assets/Scripts/TileGenerator.cs
@@ -36,15 +36,6 @@ public class TileGenerator : MonoBehaviour
             Generate();
         }
     }
-
-    public void Reset()
-    {
-        SoundManager.Instance.PlayDisplaySound();
-        BoardCheck.score = 0;
-        BoardCheck.gameover = false;
-        SceneManager.LoadScene("MainScene");
-    }
-
     public void Reroll(bool isOn)
     {
         SoundManager.Instance.PlayDisplaySound();
@@ -107,17 +98,17 @@ public class TileGenerator : MonoBehaviour
 
     private int GetRandNum()
     {
-        float randomValue = Random.Range(0f, 100f); // 0~100 »çÀÌÀÇ ·£´ı ¼ıÀÚ
+        float randomValue = Random.Range(0f, 100f); // 0~100 ?¬ì´???œë¤ ?«ì
 
-        if (randomValue < 95f)        // 95% È®·ü ¡æ 1~6
+        if (randomValue < 95f)        // 95% ?•ë¥  ??1~6
         {
             return Random.Range(1, 7);
         }
-        else if (randomValue < 99.5f) // 4.5% È®·ü ¡æ 7~10
+        else if (randomValue < 99.5f) // 4.5% ?•ë¥  ??7~10
         {
             return Random.Range(7, 11);
         }
-        else                          // 0.5% È®·ü ¡æ 11
+        else                          // 0.5% ?•ë¥  ??11
         {
             return 11;
         }
